Separate fork runtime step state recovery
The fork runtime stream module still owned streamed step-state mutation and fallback response recovery after the part-mapper split. This moves those helpers into a dedicated step-state module while re-exporting the stable functions from fork-runtime-stream so existing imports keep working.

Constraint: Preserve the public veryfront/agent export surface and keep the release metadata bumped for the PR.

Rejected: Extract the whole step runner in one pass | Step-state recovery is independently testable and keeps this release diff reviewable.

Confidence: high

Scope-risk: narrow

Directive: Keep streamed step-state mutation and fallback response recovery in fork-runtime-step-state; keep fork orchestration loops in fork-runtime-stream unless a follow-up extracts a runner.

Tested: deno test --no-check --allow-all src/agent/streaming/fork-runtime-step-state.test.ts src/agent/streaming/fork-runtime-stream.test.ts src/agent/streaming/fork-runtime-part-mapper.test.ts

Tested: deno check src/agent/streaming/fork-runtime-step-state.ts src/agent/streaming/fork-runtime-step-state.test.ts src/agent/streaming/fork-runtime-stream.ts src/agent/streaming/fork-runtime-stream.test.ts src/agent/streaming/fork-runtime-part-mapper.ts src/agent/streaming/fork-runtime-part-mapper.test.ts src/utils/version-constant.ts

Tested: git diff --check

Tested: deno task verify:quick

Not-tested: Hosted CI pending PR.
